LayoutTests/fast/forms/ios/validation-bubble-dismiss-on-tap-expected.txt

Tests that tapping the view dismisses the HTML form validation popover.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS validationBubbleContents.message is "Select this checkbox"
PASS validationBubbleContents.message is ""
PASS successfullyParsed is true

TEST COMPLETE
